+++ src/main/java/org/example/becoco/domain/post/entity/Post.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 import org.example.becoco.domain.post.entity.type.Type;
+import org.example.becoco.domain.user.domain.User;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
